Modificado metodo deleteUser de UsersService para eliminar manualmente las ofertas creadas

+++ src/main/java/es/uniovi/sdi63/sdi2223entrega163/services/UsersService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,17 +22,20 @@ public class UsersService {
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
+    @Autowired
+    private OfferService offerService;
+
     @PostConstruct
     public void init() {
     }
 
         public List<User> getUsers(){
-            List<User> users = new ArrayList<User>();
+            List<User> users = new ArrayList<>();
             usersRepository.findAll().forEach(users::add);
             return users;
         }
 
-        public User getUser(Long id){
+        public User getUser(String id){
             return usersRepository.findById(id).get();
         }
 
@@ -40,8 +44,14 @@ public class UsersService {
             usersRepository.save(user);
         }
 
-        public void deleteUser(String id){
-            usersRepository.deleteById(id);
+        public void deleteUser(String id) throws IOException {
+            var user = usersRepository.findById( id );
+            if (user.isPresent()) {
+                for (var offer : user.get().getCreatedOffers())
+                    offerService.deleteOffer( offer.getId() );
+
+                usersRepository.deleteById(id);
+            }
         }
